[Add] error handling

+++ src/controllers/UserController.ts
@@ -133,7 +133,16 @@ const deleteUser = async (req: Request, res: Response) => {
   const userId = req.header("userId");
 
   try {
-    await UserService.deleteUser(userId as string);
+    if (!userId) {
+      return res.status(statusCode.BAD_REQUEST).send(util.fail(statusCode.BAD_REQUEST, message.NULL_VALUE));
+    }
+
+    const data = await UserService.deleteUser(userId as string);
+
+    if (data === null) {
+      return res.status(statusCode.BAD_REQUEST).send(util.fail(statusCode.BAD_REQUEST, message.ALREADY_DELETED_USER));
+    }
+
     res.status(statusCode.OK).send(util.success(statusCode.OK, message.DELETE_USER_SUCCESS));
   } catch (err) {
     console.log(err);
